functions for getting locations built
It was returning the locations list for a hardcoded user on the unsecured backend get route /locations.  now trying to use the cookie with the userId, things are not working.  need to figure out how to send the cookie with the request from the front end in a way that the back end will accept.  getting some errors.